+++ src/components/Slider.vue
@@ -17,6 +17,9 @@ function getNumberOfDecimals(value) {
 function directionIsValid(direction) {
     return ['rtl', 'ltr'].indexOf(direction) > -1
 }
+function layoutIsValid(layout) {
+    return ['horizontal', 'vertical'].indexOf(layout) > -1
+}
 
 function buildOptionFormat(definition) {
     let decimals = Math.max(getNumberOfDecimals(definition.min), getNumberOfDecimals(definition.max))
@@ -53,6 +56,12 @@ function buildOptionTooltip(definition) {
         tooltips: definition.showValue && definition.modelValue !== null ? tooltipFormatter : false,
     }
 }
+function buildOptionsLayout(definition) {
+    return {
+        orientation: definition.layout,
+        direction: definition.layout === 'vertical' ? 'rtl' : definition.direction,
+    }
+}
 
 function toggleDisabled(element, disable) {
     if (disable) {
@@ -81,21 +90,21 @@ function normalizeValue(slider) {
 
 function createSlider(element, definition, context) {
     const sliderOptions = {
-        cssPrefix: 'slider--',
+        // cssPrefix: 'slider--',
         start: normalizeValue(definition),
         range: {
             min: definition.min,
             max: definition.max,
         },
         connect: [true, true],
-        direction: definition.direction,
+        ...buildOptionsLayout(definition),
         ...buildOptionStep(definition),
         ...buildOptionFormat(definition),
         ...buildOptionTooltip(definition),
     }
 
     const slider = nouislider.create(element, sliderOptions)
-    slider.on('change', function(values, handle, unencoded, tap, positions, noUiSlider) {
+    slider.on('change', values => {  /* values, handle, unencoded, tap, positions, noUiSlider */
         context.emit('update:modelValue', values[0])
     })
 
@@ -110,6 +119,7 @@ function recreateSlider(slider, element, definition, context) {
 }
 
 export default {
+    name: 'SimpleSlider',
     inheritAttrs: false,
     props: {
         min: {
@@ -143,9 +153,12 @@ export default {
         direction: {
             type: String,
             default: 'ltr',
-            validator(value) {
-                return directionIsValid(value)
-            },
+            validator: value => directionIsValid(value),
+        },
+        layout: {
+            type: String,
+            default: 'horizontal',
+            validator: value => layoutIsValid(value),
         },
     },
     setup(props, context) {
@@ -155,7 +168,7 @@ export default {
         onMounted(() => {
             slider.value = createSlider(sliderEl.value, props, context)
 
-            watch(() => props.disabled, (newValue) => toggleDisabled(slider.value.target, newValue), {
+            watch(() => props.disabled, newValue => toggleDisabled(slider.value.target, newValue), {
                 immediate: true,
             })
             watch(() => props.modelValue, newValue => {
@@ -165,6 +178,11 @@ export default {
             })
             watch(() => props.direction, newValue => {
                 if (directionIsValid(newValue)) {
+                    slider.value = recreateSlider(slider.value, sliderEl.value, props, context)
+                }
+            })
+            watch(() => props.layout, newValue => {
+                if (layoutIsValid(newValue)) {
                     slider.value = recreateSlider(slider.value, sliderEl.value, props, context)
                 }
             })
@@ -180,6 +198,7 @@ export default {
             props.min === props.modelValue && classes.push('slider--min')
             props.max === props.modelValue && classes.push('slider--max')
             props.modelValue === null && classes.push('slider--empty')
+            classes.push(`slider--layout-${props.layout}`)
 
             return classes
         })
@@ -190,152 +209,198 @@ export default {
 </script>
 
 <style lang="less">
-// Based on nouislider styles - 14.6.3
-
-// Sizes
-@sliderHeight: 20px;
-@sliderHandleSize: 30px;
-// Colors
-@sliderBackgroundColor: #FAFAFA;
-@sliderBorderColor: #D9D9D9;
-@sliderConnectBackgroundColor: #3FB8AF;
-@sliderHandleBackgroundColor: white;
-@sliderDisabledBackgroundColor: #B8B8B8;
-
+// Based on nouislider styles 15.8.0
 .slider {
-    /* Functional styling;
-     * These styles are required for noUiSlider to function.
-     * You don't need to change these rules to apply your design.
-     */
-    &--target,
-    &--target * {
-        -webkit-touch-callout: none;
-        -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
-        -webkit-user-select: none;
-        -ms-touch-action: none;
-        touch-action: none;
-        -ms-user-select: none;
-        -moz-user-select: none;
-        user-select: none;
-        -moz-box-sizing: border-box;
-        box-sizing: border-box;
-    }
-    &--target {
-        position: relative;
-    }
-    &--base,
-    &--connects {
-        width: 100%;
-        height: 100%;
-        position: relative;
-        z-index: 1;
-    }
+    --slider-border-color: #D9D9D9;
+    --slider-connect-background-color: #3FB8AF;
+    --slider-connect-border-radius: 3px;
+    --slider-handle-background-color: white;
+    --slider-disabled-background-color: #B8B8B8;
+    --slider-handle-width: 25px;
+    --slider-handle-height: 20px;
+    --slider-handle-border-radius: 3px;
+    --slider-vertical-size: 400px;
+    --slider-bar-width: 10px;
 
-    /* Wrapper for all connect elements.
-    */
-    &--connects {
-        overflow: hidden;
-        z-index: 0;
-    }
-    &--connect,
-    &--origin {
-        will-change: transform;
-        position: absolute;
-        z-index: 1;
-        top: 0;
-        right: 0;
-        -ms-transform-origin: 0 0;
-        -webkit-transform-origin: 0 0;
-        -webkit-transform-style: preserve-3d;
-        transform-origin: 0 0;
-        transform-style: flat;
-    }
-    &--connect {
-        height: 100%;
-        width: 100%;
-    }
-    &--origin {
-        height: 10%;
-        width: 10%;
-    }
-
-    /* Give origins 0 height/width so they don't interfere with clicking the
-    * connect elements.
-    */
-    &--horizontal &--origin {
-        height: 0;
-    }
-    &--handle {
-        -webkit-backface-visibility: hidden;
-        backface-visibility: hidden;
-        position: absolute;
-    }
-    &--touch-area {
-        height: 100%;
-        width: 100%;
-    }
-    &--state-tap &--connect,
-    &--state-tap &--origin {
-        -webkit-transition: transform 0.3s;
-        transition: transform 0.3s;
-    }
-    &--state-drag * {
-        cursor: inherit !important;
-    }
-
-    /* Styling;
-    * Giving the connect element a border radius causes issues with using transform: scale
-    */
-    &--target {
-        height: @sliderHeight;
-        background: @sliderBackgroundColor;
-        border-radius: 4px;
-        border: 1px solid #D3D3D3;
-    }
-    &--connects {
-        border-radius: 3px;
-    }
-    &--connect {
-        background: @sliderConnectBackgroundColor;
-    }
-
-    &--handle {
-        width: @sliderHandleSize;
-        height: @sliderHandleSize;
-        right: -@sliderHandleSize / 2;
-        top: -(@sliderHandleSize - @sliderHeight) / 2;
-        border: 1px solid @sliderBorderColor;
-        border-radius: @sliderHandleSize / 2;
-        background: @sliderHandleBackgroundColor;
-        cursor: e-resize;
-
-        &:focus {
-            box-shadow: 0 0 0 1px black;
-            outline: none;
+    &--layout {
+        &-vertical {
+            height: var(--slider-vertical-size);
         }
     }
 
-    &--tooltip {
-        display: block;
-        position: absolute;
-        border: 1px solid @sliderBorderColor;
-        border-radius: 3px;
-        padding: 5px;
-        text-align: center;
-        white-space: nowrap;
-        -webkit-transform: translate(-50%, 0);
-        transform: translate(-50%, 0);
-        left: 50%;
-        bottom: 120%;
+    .noUi {
+        &-target,
+        &-target * {
+            -webkit-touch-callout: none;
+            -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+            -webkit-user-select: none;
+            -ms-touch-action: none;
+            touch-action: none;
+            -ms-user-select: none;
+            -moz-user-select: none;
+            user-select: none;
+            -moz-box-sizing: border-box;
+            box-sizing: border-box;
+        }
+        &-target {
+            position: relative;
+        }
+
+        &-base,
+        &-connects {
+            width: 100%;
+            height: 100%;
+            position: relative;
+            z-index: 1;
+        }
+        &-connects {
+            overflow: hidden;
+            z-index: 0;
+        }
+        &-connect,
+        &-origin {
+            will-change: transform;
+            position: absolute;
+            z-index: 1;
+            top: 0;
+            right: 0;
+            height: 100%;
+            width: 100%;
+            -ms-transform-origin: 0 0;
+            -webkit-transform-origin: 0 0;
+            -webkit-transform-style: preserve-3d;
+            transform-origin: 0 0;
+            transform-style: flat;
+        }
+        // Offset direction
+        &-txt-dir-rtl {
+            .noUi-horizontal {
+                .noUi-origin {
+                    left: 0;
+                    right: auto;
+                }
+                .noUi-handle {
+                    // left: -17px;
+                    right: auto;
+                }
+            }
+        }
+        // Give origins 0 height/width so they don't interfere with clicking the connect elements.
+        &-vertical {
+            width: var(--slider-bar-width);
+            height: 100%;
+
+            .noUi-origin {
+                top: calc(-100% + 1.5 * var(--slider-handle-width));
+                height: calc(100% - var(--slider-handle-width));
+                left: 50%;
+                width: 0;
+            }
+            .noUi-handle {
+                height: var(--slider-handle-width);
+                width: var(--slider-handle-height);
+                transform: translate(50%, 50%);
+                bottom: 0;
+            }
+            .noUi-tooltip {
+                -webkit-transform: translate(0, -50%);
+                transform: translate(0, -50%);
+                top: 50%;
+                left: 120%;
+            }
+        }
+        &-horizontal {
+            height: var(--slider-bar-width);
+
+            .noUi-origin {
+                width: auto;
+                height: 0;
+                top: 50%;
+                left: calc(var(--slider-handle-width) / 2);
+                right: calc(var(--slider-handle-width) / 2);
+            }
+            .noUi-handle {
+                width: var(--slider-handle-width);
+                height: var(--slider-handle-height);
+                transform: translate(50%, -50%);
+                top: 0;
+            }
+            .noUi-tooltip {
+                -webkit-transform: translate(-50%, 0);
+                transform: translate(-50%, 0);
+                left: 50%;
+                bottom: 120%;
+            }
+        }
+
+        &-handle {
+            -webkit-backface-visibility: hidden;
+            backface-visibility: hidden;
+            position: absolute;
+            right: 0;
+
+            cursor: grab;
+            border: 1px solid var(--slider-border-color);
+            border-radius: var(--slider-handle-border-radius);
+            background-color: var(--slider-handle-background-color);
+            box-shadow: 0 3px 6px -3px silver;
+
+            &.noUi-active {
+                cursor: grabbing;
+            }
+        }
+
+        &-touch-area {
+            height: 100%;
+            width: 100%;
+        }
+
+        &-state-tap {
+            .noUi-connect,
+            .noUi-origin {
+                -webkit-transition: transform 0.3s;
+                transition: transform 0.3s;
+            }
+        }
+
+        &-state-drag * {
+            cursor: inherit !important;
+        }
+
+        &-target {
+            background: var(--slider-handle-background-color);
+            border-radius: calc(var(--slider-connect-border-radius) + 1px);
+            border: 1px solid var(--slider-border-color);
+        }
+
+        &-connects {
+            border-radius: var(--slider-connect-border-radius);
+        }
+        &-connect {
+            background: var(--slider-connect-background-color);
+        }
+
+        &-tooltip {
+            display: block;
+            position: absolute;
+            border: 1px solid var(--slider-border-color);
+            border-radius: var(--slider-handle-border-radius);
+            background: var(--slider-handle-background-color);
+            padding: 3px 5px;
+            text-align: center;
+            white-space: nowrap;
+        }
     }
 
-    [disabled] &--connect {
-        background: @sliderDisabledBackgroundColor;
+    &--disabled {
+        .noUi {
+            &-connect {
+                background: var(--slider-disabled-background-color);
+            }
+            &-target,
+            &-handle {
+                cursor: not-allowed;
+            }
+        }
     }
-    [disabled]&--target,
-    [disabled]&--handle,
-    [disabled] &--handle {
-        cursor: not-allowed;
-    }
-}
-</style>
+}</style>